Fix???
Note this was done with a JS editor in the browser!!!
Main problem was the rightGuess[index] is not working properly (or at least as far as I can tell). 
Also I don't know why "let letterPosition = rightGuess.indexOf(currentGuess[i])" was used when you literally already have the correct index (i).